requirements.txt
Flask==2.1.0
flask-cors==3.0.10
pymongo==3.12.0
python-dotenv==0.19.1
